==> Code/2_MeSH_ID_Extraction.ipynb <==
import pandas as pd
from sentence_transformers import SentenceTransformer
import numpy as np
import json
import torch

device = "cuda" if torch.cuda.is_available() else "cpu"
print(f"Using device: {device}")

model = SentenceTransformer("abhinand/MedEmbed-base-v0.1", device=device)

d1 = pd.read_csv("/kaggle/input/descriptor/descriptor-terms.tsv", sep="\t")

terms = d1['TermName'].tolist()
term_ids = d1['DescriptorUI'].tolist()

df=pd.read_excel("/kaggle/input/descriptor/usecase_4_.xlsx")

condition=pd.DataFrame(df["Conditions"].unique())
conditions=condition[0].to_list()


term_embeddings = model.encode(terms, device=device)
condition_embeddings = model.encode(conditions, device=device)

def find_top_k_matches(condition_embedding, term_embeddings, terms, term_ids, k=3):
    similarities = np.dot(term_embeddings, condition_embedding) / (
        np.linalg.norm(term_embeddings, axis=1) * np.linalg.norm(condition_embedding)
    )

    top_k_indices = np.argsort(similarities)[-k:][::-1]

    top_k_matches = []
    for idx in top_k_indices:
        top_k_matches.append({
            "TermID": term_ids[idx],
            "TermName": terms[idx],
            "Similarity": float(similarities[idx])
        })
    return top_k_matches


from tqdm import tqdm

k = 5
result = {}
for condition, condition_embedding in tqdm(zip(conditions, condition_embeddings), total=len(conditions), desc="Processing conditions"):
    top_k_matches = find_top_k_matches(condition_embedding, term_embeddings, terms, term_ids, k)
    result[condition] = top_k_matches


json_result = json.dumps(result, indent=4)

output_file = "/kaggle/working/output.json"
with open(output_file, "w") as f:
    json.dump(result, f, indent=4)

print(f"JSON result saved to {output_file}")
