==> Code/5_Competition_extraction.ipynb <==
import pandas as pd
import numpy as np

df_cities = pd.read_csv('/content/drive/MyDrive/NOVARTIS NEST - BioLogits/Data/Population_extracted.csv')
df_actual = pd.read_excel('/content/drive/MyDrive/NOVARTIS NEST - BioLogits/Data/usecase_4_.xlsx')
df_cities.shape, df_actual.shape

df_cities.columns

df = pd.merge(df_cities, df_actual, on='NCT Number', how='inner')
df.shape

df.columns

df = df.rename(columns={'Unnamed: 0': 'index'})

exploded = df.explode('Cities')

def get_overlapping_indices(group):
    starts = group['Start Date'].values
    ends = group['Completion Date'].values

    overlap_matrix = (starts[:, None] <= ends) & (starts <= ends[:, None]).T
    np.fill_diagonal(overlap_matrix, False)

    return pd.Series(
        [set(group['index'][overlap_matrix[i]].values) for i in range(len(group))],
        index=group.index
    )

exploded['overlaps'] = exploded.groupby('Cities', group_keys=False).apply(get_overlapping_indices)


df = df.merge(
    exploded.groupby('index')['overlaps'].agg(lambda x: set().union(*x)),
    left_on='index',
    right_index=True
)
df['competition'] = df['overlaps'].apply(len)
df = df.drop(columns=['index', 'overlaps'])

print(df[['Cities', 'Start Date', 'Completion Date', 'competition']])

df_check = df[df['Cities']=="['Pittsburgh']"]
df_check = df_check[['Cities', 'competition', 'Start Date', 'Completion Date']]
df_check

df = df[['NCT Number','competition']]
df

df.to_csv('/content/drive/MyDrive/NOVARTIS NEST - BioLogits/Data/competition.csv', index=False)

## time

df_actual['Start Date'] = pd.to_datetime(df_actual['Start Date'])
df_actual['Completion Date'] = pd.to_datetime(df_actual['Completion Date'])

# Calculate duration in days
df_actual['duration'] = (df_actual['Completion Date'] - df_actual['Start Date']).dt.days

df_actual[['Start Date', 'Completion Date', 'duration']]

df_actual = df_actual[['NCT Number', 'duration']]
df_actual

df_actual.to_csv('/content/drive/MyDrive/NOVARTIS NEST - BioLogits/Data/duration.csv', index=False)
