==> Code/6_Data_Merging.ipynb <==
import pandas as pd

base_df = pd.read_csv("/content/drive/MyDrive/NOVARTIS NEST - BioLogits/Data/usecase_4_ec_added_entropy.csv")

base_df.columns

drop_cols = ['Unnamed: 0', 'Study Title', 'Study URL',
       'Brief Summary', 'Conditions', 'Interventions',
       'Primary Outcome Measures', 'Secondary Outcome Measures',
       'Other Outcome Measures', 'Sponsor', 'Collaborators',
             'Other IDs', 'Start Date', 'Primary Completion Date', 'Completion Date',
       'First Posted', 'Results First Posted', 'Last Update Posted',
       'Locations', 'Eligibility Criteria', 'Study Design', 'Study Type'
             ]

base_df.drop(drop_cols, axis=1, inplace=True)

base_df

spons_df = pd.read_csv("/content/drive/MyDrive/NOVARTIS NEST - BioLogits/Data/usecase_4_sponsor_success_added_with_nan.csv")

spons_df.columns

base_df['sponsor_success_rate'] = spons_df["sponsor_success_rate"]

prev_disease_data = pd.read_csv('/content/drive/MyDrive/NOVARTIS NEST - BioLogits/Data/usecase_4_prevalent_disease_added.csv')

prev_disease_data.columns

base_df['prevalent_disease'] = prev_disease_data['Prevalence_disease']
base_df['mesh_id'] = prev_disease_data['mesh_id']

popu_df = pd.read_csv('/content/drive/MyDrive/NOVARTIS NEST - BioLogits/Data/Population_extracted.csv')

popu_df.columns

base_df['no_of_sites'] = popu_df['no_of_sites']
base_df['population'] = popu_df['population']

comp_df = pd.read_csv('/content/drive/MyDrive/NOVARTIS NEST - BioLogits/Data/competition.csv')

comp_df.columns

base_df['competition'] = comp_df['competition']

study_des = pd.read_csv('/content/drive/MyDrive/NOVARTIS NEST - BioLogits/Data/after_study_design_new.csv')

study_des.columns

drop_here = ['Study Title', 'Study URL', 'Study Status',
       'Brief Summary', 'Study Results', 'Conditions', 'Interventions',
       'Primary Outcome Measures', 'Secondary Outcome Measures',
       'Other Outcome Measures', 'Sponsor', 'Collaborators', 'Sex', 'Age',
       'Enrollment', 'Funder Type', 'Study Type', 'Other IDs', 'Start Date',
       'Primary Completion Date', 'Completion Date', 'First Posted',
       'Results First Posted', 'Last Update Posted', 'Locations',
       'Study Recruitment Rate']

study_des.drop(drop_here, axis=1, inplace=True)

study_des.columns

# merge base_df and study_des on 'NCT Number' column
base_df = pd.merge(base_df, study_des, on='NCT Number', how='left')

base_df.columns

dura_df = pd.read_csv('/content/drive/MyDrive/NOVARTIS NEST - BioLogits/Data/duration.csv')

dura_df.columns

base_df['study_duration'] = dura_df['duration']

base_df.columns

one_hot_cols = ['Study Status', 'Study Results', 'Sex', 'Funder Type']

final_df = pd.get_dummies(base_df, columns=one_hot_cols, dtype=int)


final_df.columns

final_df.drop('Phases', axis=1, inplace=True)

final_df.columns

final_df['Enrollment'].sort_values()

# remove the row with max Enrollment value
final_df = final_df[final_df['Enrollment'] != final_df['Enrollment'].max()]

final_df['Enrollment'].sort_values()

final_df.columns

nan_list = ['Allocation_NA', 'Allocation_NON_RANDOMIZED',
       'Allocation_RANDOMIZED', 'Intervention Model_CROSSOVER',
       'Intervention Model_PARALLEL', 'Intervention Model_SEQUENTIAL',
       'Intervention Model_SINGLE_GROUP', 'Masking Complexity_DOUBLE',
       'Masking Complexity_NONE', 'Masking Complexity_QUADRUPLE',
       'Masking Complexity_SINGLE', 'Masking Complexity_TRIPLE',
       'Primary Purpose_BASIC_SCIENCE', 'Primary Purpose_OTHER',
       'Primary Purpose_PREVENTION', 'Primary Purpose_TREATMENT']

final_df[nan_list] = final_df[nan_list].fillna(0)

columns_to_fill = ['sponsor_success_rate', 'prevalent_disease', 'population']
final_df[columns_to_fill] = final_df[columns_to_fill].apply(lambda col: col.fillna(col.mean()))

final_df.isna().sum()

final_df.columns

final_df.describe()

numerical_cols = ['Enrollment', 'ec_entropy_1st', 'ec_entropy_2nd', 'sponsor_success_rate',
       'prevalent_disease', 'no_of_sites', 'population',
       'competition']

# Apply Min-Max Scaling
final_df[numerical_cols] = final_df[numerical_cols].apply(lambda col: (col - col.min()) / (col.max() - col.min()))

final_df.describe()

final_df['mesh_id'] = final_df['mesh_id'].apply(lambda x: eval(x))

freq_dict = {}
for sub_list in final_df['mesh_id']:
  for item in sub_list:
    if item in freq_dict:
      freq_dict[item] += 1
    else:
      freq_dict[item] = 1


# Frequency encoding: Replace each list with the sum of frequencies of its elements
for index, row in final_df.iterrows():
    freq_sum = sum(freq_dict[item] for item in row['mesh_id'])/len(row['mesh_id'])
    final_df.at[index, 'mesh_id_encoded'] = freq_sum

final_df['mesh_id_encoded']

numerical_cols = ['mesh_id_encoded']

# Apply Min-Max Scaling
final_df[numerical_cols] = final_df[numerical_cols].apply(lambda col: (col - col.min()) / (col.max() - col.min()))

final_dataset = final_df.drop('mesh_id', axis=1)

one_hot_cols = ['Age']

final_dataset = pd.get_dummies(final_dataset, columns=one_hot_cols, dtype=int)


final_dataset.to_csv('/content/drive/MyDrive/NOVARTIS NEST - BioLogits/Data/final_final_dataset.csv', index=False)
